chore: move scoreboard and its interface to service directory
