+++ scoreboard/src/main/java/com/sportradar/scoreboard/service/ScoreboardService.java
@@ -1,4 +1,4 @@
-package com.sportradar.scoreboard;
+package com.sportradar.scoreboard.service;
 
 public interface ScoreboardService {
     void startMatch(String homeTeam, String awayTeam);
